refactor(app): extract homepage migration effect to layout-homepage-migration.ts
Slice 12 of #1056 layout governance line — pure extraction.

Moves the fire-and-forget v7 homepage-draft migration createEffect out of
pages/layout.tsx into a new useHomepageMigration({ currentDir, platform })
factory, mirroring the sibling useUpdatePolling / layout-update-polling.ts
injection pattern. The effect registers in the parent reactive root exactly
as before (call sited where the effect previously lived), so owner and
registration order are unchanged.

Five now-unused imports drop from layout.tsx (runHomepageMigration,
HOMEPAGE_MIGRATION_SENTINEL_KEY, LegacyHomepagePromptStore, usePortableDraft,
createMigrationStorageIO); the new file imports them directly. Persist and
createEffect stay in layout.tsx (still used elsewhere).

No behaviour / DOM / aria / copy / storage-key change — the effect body,
sentinel handling, idempotent guard, and the desktop-vs-web storage branch
(which already lived in homepage-migration-storage.ts, not layout.tsx) are
byte-for-byte identical.

layout.tsx 1084 -> 1026 (-58).

## Verification

- bun run typecheck clean
- Source-guard tests (shell-frame-contract, update-install-flow-source) +
  homepage-migration logic tests: 27 pass (the guards make only negative
  assertions unrelated to migration, so none needed to move)
- Full bun run test:unit: 1734 pass / 0 fail (no regression)

No new unit test: the extracted unit is a pure effect wrapper with no memo /
pure-helper paths, and bun's solid-js server build no-ops createEffect so the
effect body is not drivable under bun (same constraint as slice 10). The
migration logic it wires is already covered by homepage-migration.test.ts;
equivalence of the move is verified by codex review.
